Mnogo izmena modela i povezivanje sa backom

+++ src/Front/app/src/main/java/com/example/front/model/DTO/CheckAvailabilityReqDTO.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class CheckAvailabilityReqDTO (
     @SerializedName("id") val id: Int,
     @SerializedName("size") val size: String = "None",
-    @SerializedName("quantity") val quantity: Double,
+    @SerializedName("quantity") val quantity: Int,
 )
