+++ src/Front/app/src/main/java/com/example/front/model/product/ProductInCart.kt
@@ -6,12 +6,12 @@ class ProductInCart : RealmObject {
 //    @PrimaryKey
     var id: Int = 0
     var name: String = ""
-    var price: Float = 0.0f
-    var quantity: Double = 0.0
+    var price: Double = 0.0
+    var quantity: Int = 0
     var shopId: Int = 0
     var shopName: String = ""
     var image: String = ""
     var metric: String = ""
     var size: String = "None"
-    var available: Double = 0.0
+    var available: Int = 0
 }
